* mpz/bin_uiui.c: Fix result for n==0 and n==k.
This is per a bug-gmp report by Hans J Lahne <hjlas@online.no>, but
not using the patch offerred, instead taking the opportunity to
flatten out the nested conditionals.

The diff for this revision is a bit horrible, but only the initial
setups have changed, the main loop has just been reindented.