* mpn/generic/gcd.c: Inline two small mpn_bdivmod calls, use 	MPN_COPY_INCR not MPN_COPY in one place.
This inlining and the use of modlimb_invert is a speedup on K7 by
about a factor of 1.4x on small sizes (up to 40 limbs say), and on K6
by about 1.2x.  modlimb_invert accounts for most of this.  Large sizes
aren't sped-up, since they're dominated by addmul_1 etc.

GCD_ACCEL_THRESHOLD is much reduced on K7 too, from about 20 limbs to
about 5 (the latter being roughly the hoped-for value).