* mpz/jacobi.c: Don't retaining old behaviour of mpz_jacobi on even b 	(it wasn't documented in 3.1.1). 	* mpz/jacobi.c, gmp.h (mpz_kronecker, mpz_legendre): Remove separate 	entrypoints, just #define to mpz_jacobi.
The old behaviour of ignoring factors of 2 on b was never documented
and doesn't seem particularly useful, so there's really no need to
keep it, better just make one kronecker style function and share it.

With respect to gmp 3.1.1 we've now got mpz_jacobi fixed to allow
negative b, and mpz_kronecker added.

Could have just told applications mpz_jacobi now has the kronecker
extension, but it matches mpz_kronecker_ui etc better if there's an
mpz_kronecker alias.

__gmpz_jacobi is kept as the primary library entrypoint and
__gmpz_legendre just for compatibility.  Done this way around on the
assumption existing binaries are more likely to be using mpz_jacobi
than mpz_legendre.  Obviously a recompile moves it to the right one if
this is not so.